refactor(topology): move interface PartitionDistributor from raft to topology
The concrete implementations are provided in broker.
