+++ broker/src/test/java/io/camunda/zeebe/broker/partitioning/RaftRolesTest.java
@@ -21,7 +21,7 @@ import io.atomix.raft.RaftServer.Role;
 import io.atomix.raft.partition.RaftPartition;
 import io.atomix.raft.partition.RaftPartitionConfig;
 import io.atomix.raft.partition.RaftStorageConfig;
-import io.atomix.raft.partition.RoundRobinPartitionDistributor;
+import io.camunda.zeebe.broker.partitioning.distribution.RoundRobinPartitionDistributor;
 import java.io.File;
 import java.util.ArrayList;
 import java.util.Collections;
